requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
bidict==0.22.0
blinker==1.9.0
cachelib==0.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.3
cysystemd==2.0.1
Flask==3.1.0
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SocketIO==5.3.4
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib-metadata==4.12.0
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.1.0
packaging==24.2
paho-mqtt==2.1.0
propcache==0.2.1
pyftdi==0.54.0
pyserial==3.5
pyTelegramBotAPI==4.26.0
python-engineio==4.4.1
python-socketio==5.8.0
pyusb==1.2.1
PyYAML==6.0.2
requests==2.32.3
simple-websocket==1.1.0
six==1.11.0
systemd==0.17.1
systemd-python==235
sysv_ipc==1.1.0
typing_extensions==4.6.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zipp==3.8.1
